+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [#266](https://github.com/strongqa/howitzer/issues/266) ruby 2.2 no longer supported
 - [#265](https://github.com/strongqa/howitzer/issues/265) Add Appium driver
 ### Bug-fixes
+- [#274](https://github.com/strongqa/howitzer/issues/274) Stopped PhantomJS supporting
 
 ## [v2.1.1](https://github.com/strongqa/howitzer/compare/v2.1.0...v2.1.1)